requirements.txt
aiogram==2.12
aiohttp==3.7.4.post0
aiosqlite==0.16.1
async-property==0.2.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
FlightRadarAPI==1.1.5
idna==2.10
iso8601==0.1.14
marshmallow==3.10.0
multidict==5.1.0
PyPika==0.44.1
pytz==2020.5
requests==2.25.1
soupsieve==2.2
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
